Update and rename add two no.py to add_two_nums.py
Cleaned and Optimized the program for better readability and execution.
